requirements.txt
requests
PyYAML
urllib3

